Extract create_conversation into commons file
